requirements.txt
pyaudio
numpy
openai-whisper